refactor(sdl): Split parser.rs into smaller modules
Split the large parser.rs file into focused modules:
- warnings.rs (132 lines): ParseWarning, DirectiveLocation, ParseOutput
- directives.rs (130 lines): Known directives, argument helpers
- parser.rs (809 lines non-test): SdlParser and core logic

This follows CLAUDE.md guidelines for file organization and improves
maintainability by separating concerns.

Closes #34
