+++ crates/query/src/sdl_parse/mod.rs
@@ -3,9 +3,10 @@
 //! This module parses GraphQL Schema Definition Language (SDL) and converts
 //! type definitions into DefraDB CollectionVersion schemas.
 
+mod directives;
 mod parser;
+mod warnings;
 
-pub use parser::{
-    parse_sdl, parse_sdl_with_warnings, DirectiveLocation, ParseOutput, ParseWarning,
-    ParsedDirectives, SdlParser,
-};
+pub use directives::ParsedDirectives;
+pub use parser::{parse_sdl, parse_sdl_with_warnings, SdlParser};
+pub use warnings::{DirectiveLocation, ParseOutput, ParseWarning};
