+++ crates/query/src/sdl_parse/parser.rs
@@ -14,209 +14,12 @@ use schema::{
 use sha2::{Digest, Sha256};
 use std::collections::HashMap;
 
-/// Warnings generated during SDL parsing
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum ParseWarning {
-    /// Unknown directive encountered (forward compatibility)
-    UnknownDirective {
-        directive_name: String,
-        location: DirectiveLocation,
-        type_name: String,
-        field_name: Option<String>,
-    },
-    /// Unknown argument on a known directive
-    UnknownDirectiveArgument {
-        directive_name: String,
-        argument_name: String,
-        type_name: String,
-        field_name: Option<String>,
-    },
-    /// Directive is recognized but not yet implemented
-    UnimplementedDirective {
-        directive_name: String,
-        type_name: String,
-        field_name: Option<String>,
-    },
-    /// Argument has wrong type (e.g., string instead of bool)
-    InvalidArgumentType {
-        directive_name: String,
-        argument_name: String,
-        expected_type: String,
-        type_name: String,
-        field_name: Option<String>,
-    },
-}
-
-impl std::fmt::Display for ParseWarning {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            ParseWarning::UnknownDirective {
-                directive_name,
-                location,
-                type_name,
-                field_name,
-            } => {
-                let location_str = match location {
-                    DirectiveLocation::Type => format!("type {}", type_name),
-                    DirectiveLocation::Field => {
-                        format!(
-                            "field {}.{}",
-                            type_name,
-                            field_name.as_deref().unwrap_or("?")
-                        )
-                    }
-                };
-                write!(
-                    f,
-                    "unknown directive @{} on {} (ignored for forward compatibility)",
-                    directive_name, location_str
-                )
-            }
-            ParseWarning::UnknownDirectiveArgument {
-                directive_name,
-                argument_name,
-                type_name,
-                field_name,
-            } => {
-                let location_str = format_location(type_name, field_name.as_deref());
-                write!(
-                    f,
-                    "unknown argument '{}' on directive @{} at {}",
-                    argument_name, directive_name, location_str
-                )
-            }
-            ParseWarning::UnimplementedDirective {
-                directive_name,
-                type_name,
-                field_name,
-            } => {
-                let location_str = format_location(type_name, field_name.as_deref());
-                write!(
-                    f,
-                    "directive @{} at {} is recognized but not yet implemented",
-                    directive_name, location_str
-                )
-            }
-            ParseWarning::InvalidArgumentType {
-                directive_name,
-                argument_name,
-                expected_type,
-                type_name,
-                field_name,
-            } => {
-                let location_str = format_location(type_name, field_name.as_deref());
-                write!(
-                    f,
-                    "argument '{}' on @{} at {} should be {}, value ignored",
-                    argument_name, directive_name, location_str, expected_type
-                )
-            }
-        }
-    }
-}
-
-/// Format a location string for warnings
-fn format_location(type_name: &str, field_name: Option<&str>) -> String {
-    if let Some(fname) = field_name {
-        format!("{}.{}", type_name, fname)
-    } else {
-        type_name.to_string()
-    }
-}
-
-/// Location where a directive was found
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum DirectiveLocation {
-    Type,
-    Field,
-}
-
-/// Result of SDL parsing including warnings
-#[derive(Debug)]
-pub struct ParseOutput {
-    /// Parsed collection versions
-    pub collections: Vec<CollectionVersion>,
-    /// Warnings generated during parsing
-    pub warnings: Vec<ParseWarning>,
-}
-
-/// Known field-level directives
-const KNOWN_FIELD_DIRECTIVES: &[&str] = &[
-    "primary",
-    "crdt",
-    "index",
-    "relation",
-    "default",
-    "constraints",
-    "embedding",
-    "encryptedIndex",
-    "policy", // Go allows @policy on fields in some contexts
-];
-
-/// Known type-level directives
-const KNOWN_TYPE_DIRECTIVES: &[&str] = &["index", "materialized", "branchable", "policy"];
-
-/// Known arguments for each directive
-fn known_directive_arguments(directive_name: &str) -> &'static [&'static str] {
-    match directive_name {
-        "primary" => &[],
-        "crdt" => &["type"],
-        "index" => &["name", "unique", "direction", "fields", "includes"],
-        "relation" => &["name"],
-        "default" => &[
-            "string", "value", "bool", "int", "float", "float32", "float64", "dateTime", "json",
-            "blob",
-        ],
-        "constraints" => &["size"],
-        "materialized" => &["if"],
-        "branchable" => &["if"],
-        "policy" => &["id", "resource"],
-        "embedding" => &["provider", "model", "url", "fields", "template"],
-        "encryptedIndex" => &["type"],
-        _ => &[],
-    }
-}
-
-// =============================================================================
-// Directive argument helpers (standalone functions)
-// =============================================================================
-
-/// Find a directive argument by name
-fn get_directive_arg<'a, 'b>(
-    directive: &'a Directive<'b, String>,
-    arg_name: &str,
-) -> Option<&'a graphql_parser::schema::Value<'b, String>> {
-    directive
-        .arguments
-        .iter()
-        .find(|(name, _)| name == arg_name)
-        .map(|(_, value)| value)
-}
-
-/// Get a string argument from a directive
-fn get_directive_string(directive: &Directive<'_, String>, arg_name: &str) -> Option<String> {
-    match get_directive_arg(directive, arg_name)? {
-        graphql_parser::schema::Value::String(s) | graphql_parser::schema::Value::Enum(s) => {
-            Some(s.clone())
-        }
-        _ => None,
-    }
-}
-
-/// Get a string list argument from a directive
-fn get_directive_string_list(directive: &Directive<'_, String>, arg_name: &str) -> Vec<String> {
-    match get_directive_arg(directive, arg_name) {
-        Some(graphql_parser::schema::Value::List(items)) => items
-            .iter()
-            .filter_map(|v| match v {
-                graphql_parser::schema::Value::String(s)
-                | graphql_parser::schema::Value::Enum(s) => Some(s.clone()),
-                _ => None,
-            })
-            .collect(),
-        _ => Vec::new(),
-    }
-}
+use super::directives::{
+    default_type_error, get_directive_arg, get_directive_string, get_directive_string_list,
+    known_directive_arguments, IndexConfig, IndexDirection, ParsedDirectives,
+    KNOWN_FIELD_DIRECTIVES, KNOWN_TYPE_DIRECTIVES,
+};
+use super::warnings::{DirectiveLocation, ParseOutput, ParseWarning};
 
 /// Parse @policy directive arguments
 fn parse_policy_directive(directive: &Directive<'_, String>) -> Result<PolicyConfig> {
@@ -227,50 +30,6 @@ fn parse_policy_directive(directive: &Directive<'_, String>) -> Result<PolicyCon
         .ok_or_else(|| QueryError::parse("@policy directive requires 'resource' argument"))?;
 
     Ok(PolicyConfig { id, resource })
-}
-
-/// Create a type mismatch error for @default directive
-fn default_type_error(
-    arg_name: &str,
-    expected: &str,
-    actual: &graphql_parser::schema::Value<'_, String>,
-) -> QueryError {
-    QueryError::parse(format!(
-        "@default '{}' argument must be a {}, got {:?}",
-        arg_name, expected, actual
-    ))
-}
-
-/// Parsed directive information from a field
-#[derive(Debug, Default, Clone)]
-pub struct ParsedDirectives {
-    /// Whether this field is the primary side of a relation
-    pub is_primary: bool,
-    /// CRDT type override (default is LwwRegister)
-    pub crdt_type: Option<CType>,
-    /// Index configuration
-    pub index: Option<IndexConfig>,
-    /// Explicit relation name from @relation directive
-    pub relation_name: Option<String>,
-    /// Default value from @default directive
-    pub default_value: Option<serde_json::Value>,
-    /// Array size constraint from @constraints directive
-    pub size_constraint: Option<usize>,
-}
-
-/// Index configuration from @index directive
-#[derive(Debug, Clone)]
-pub struct IndexConfig {
-    pub name: Option<String>,
-    pub unique: bool,
-    pub direction: IndexDirection,
-}
-
-#[derive(Debug, Clone, Copy, Default)]
-pub enum IndexDirection {
-    #[default]
-    Asc,
-    Desc,
 }
 
 /// SDL parser for DefraDB schemas
